completed offline connection
