--- app/src/main/java/com/example/zoomwroom/FragmentDriverAccepted.java
@@ -1,104 +1,0 @@
-package com.example.zoomwroom;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.TextView;
-
-import com.example.zoomwroom.Entities.DriveRequest;
-import com.example.zoomwroom.database.MyDataBase;
-
-
-
-public class FragmentDriverAccepted extends FragmentCreateRide {
-    public void FragmentDriverAccepted(){};
-    private Button cancel;
-    private Button complete;
-    private Button confirm;
-
-    @Override
-    public void onCreate(Bundle savedInstancesState) {
-        super.onCreate(savedInstancesState);
-
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-
-        // ******************** Setting textview to the appropriate values ***************
-
-        View view = inflater.inflate(R.layout.fragment_driver_accepted, container, false);
-        Bundle bundle = getArguments();
-        DriveRequest driveRequest = (DriveRequest) bundle.getSerializable("driveRequest");
-
-        TextView driverName = view.findViewById(R.id.driver_name);
-        String stringName = MyDataBase.getDriver(driveRequest.getDriverID()).getName();
-        driverName.setText(stringName);
-
-        TextView driverUserName = view.findViewById(R.id.driver_username);
-        String stringUsername = MyDataBase.getDriver(driveRequest.getDriverID()).getUserName();
-        driverUserName.setText(stringUsername);
-
-        TextView destination = view.findViewById(R.id.destination_text);
-        String stringDestination =  Double.toString(driveRequest.getDestinationLng()) + ", " + Double.toString(driveRequest.getDestinationLat());
-        destination.setText("Destination: " + stringDestination);
-
-        TextView pickup = view.findViewById(R.id.pickup_text);
-        String stringPickup = Double.toString(driveRequest.getPickupLocationLng()) + ", " + Double.toString(driveRequest.getPickupLocationLat());
-        pickup.setText("Pickup: " + stringPickup);
-
-        TextView fare = view.findViewById(R.id.fare_text);
-        String stringFare = Float.toString(driveRequest.getSuggestedFare());
-        fare.setText("Fare: " + stringFare);
-
-        // ******************** Setting textview to the appropriate values ***************
-
-        TextView rideStatus = ((RiderHomeActivity) getActivity()).findViewById(R.id.rideStatus);
-        rideStatus.setText("ACCEPTED BY DRIVER");
-
-        // Cancel button
-        // Set the status to 5, then update firebase
-        // restart main activity
-        cancel = view.findViewById(R.id.cancel_button);
-        cancel.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                driveRequest.setStatus(5);
-                MyDataBase.updateRequest(driveRequest);
-                Intent intent = new Intent(getActivity(), RiderHomeActivity.class);
-                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                startActivity(intent);
-            }
-        });
-
-        // hide the button
-        // ride is now confirmed
-        confirm = view.findViewById(R.id.confirm_button);
-        confirm.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                driveRequest.setStatus(2);
-                MyDataBase.updateRequest(driveRequest);
-                rideStatus.setText("WAITING FOR DRIVER");
-                confirm.setVisibility(View.GONE);
-            }
-        });
-        complete = view.findViewById(R.id.complete_button);
-
-
-
-
-
-
-        return view;
-    }
-
-    public void phaseThree(){
-        complete.setVisibility(View.VISIBLE);
-        cancel.setVisibility(View.INVISIBLE);
-    }
-
-}
