+++ app/src/main/res/layout/fragment_create_ride.xml
@@ -102,7 +102,8 @@
             android:layout_weight="1"
             android:textColor="@color/colorLightGray"
             android:textSize="15sp"
-            android:visibility="invisible" />
+            android:visibility="invisible"
+            android:clickable="true"/>
 
 
     </LinearLayout>
