--- app/src/main/res/layout/fragment_driver_accepted.xml
@@ -1,138 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="330dp"
-    android:background="@color/colorWhite">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="220dp"
-        android:orientation="vertical">
-
-
-        <TextView
-            android:id="@+id/destination_text"
-            android:layout_width="352dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:inputType="text"
-            android:text="Destination: "
-            android:textColor="@color/colorLightGray"
-            android:textSize="15dp"
-            android:layout_marginTop="20dp"/>
-
-
-        <View
-            android:layout_width="350dp"
-            android:layout_height="1dp"
-            android:layout_marginStart="10dp"
-            android:background="@color/colorLightGray" />
-
-        <TextView
-            android:id="@+id/pickup_text"
-            android:layout_width="351dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:inputType="text"
-            android:text="Pickup: "
-            android:textColor="@color/colorLightGray"
-            android:textSize="15dp"
-            android:layout_marginTop="20dp"/>
-
-        <View
-            android:layout_width="350dp"
-            android:layout_height="1dp"
-            android:layout_marginStart="10dp"
-            android:background="@color/colorLightGray" />
-
-
-        <EditText
-            android:id="@+id/fare_text"
-            android:layout_width="351dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:text="Fare: "
-            android:textColor="@color/colorLightGray"
-            android:textSize="15sp"
-            android:layout_marginTop="20dp"/>
-
-        <View
-            android:layout_width="350dp"
-            android:layout_height="1dp"
-            android:layout_marginStart="10dp"
-            android:background="@color/colorLightGray" />
-
-        <TextView
-            android:id="@+id/driver_name"
-            android:layout_width="239dp"
-            android:layout_height="45dp"
-            android:layout_gravity="center"
-            android:layout_marginTop="15dp"
-            android:textColor="@color/colorLightGray"
-            android:textSize="30sp"></TextView>
-
-        <TextView
-            android:id="@+id/driver_username"
-            android:layout_width="239dp"
-            android:layout_height="45dp"
-            android:textColor="@color/colorLightGray"
-            android:textSize="20sp"
-            android:layout_gravity="center">
-        </TextView>
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <Button
-            android:id="@+id/confirm_button"
-            style="@style/PrimaryButtonRegisterRider"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="50dp"
-            android:backgroundTint="@color/colorAccentPurple"
-            android:text="CONFIRM"
-            android:textColorHint="@color/colorWhite">
-
-        </Button>
-
-        <Button
-            android:id="@+id/cancel_button"
-            style="@style/PrimaryButtonRegisterRider"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="50dp"
-            android:backgroundTint="@color/colorAccentPurple"
-            android:text="CANCEL"
-            android:textColorHint="@color/colorWhite">
-
-        </Button>
-
-        <Button
-            android:id="@+id/complete_button"
-            style="@style/PrimaryButtonRegisterRider"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="50dp"
-            android:backgroundTint="@color/colorAccentPurple"
-            android:text="RIDE COMPLETED"
-            android:textColorHint="@color/colorWhite"
-            android:visibility="invisible">
-        </Button>
-
-
-
-
-
-    </LinearLayout>
-
-
-
-
-
-</LinearLayout>
